+++ book_manage_service/src/main/java/com/uga/book_manage_service/service/BookRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import com.uga.book_manage_service.model.Book;
 
 @Repository
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 	
-	public Book findById(int id);
+	public Book findById(long id);
 }
